Convert all colors to HSL
The Hue Saturation Lightness system lets you define a single hue, and
then modify the saturation and lightness of the color.

In total 70 different colors are used in Critic, and they have been
chosen because they looked ok in a particular circumstance, and not
because they necessarily matched the rest of the color scheme.

Using HSL, we can select for example 2 main and 2 complementary hues,
and stick to those. If someone needs a new color for something, they
will select a hue and tweak the lightness and saturation, which means
it's easy to make sure that no one introduces a color that doesn't match
the rest of the color scheme.

HSL is also very convenient to tweak in the developer tools, where you
can just select saturation or lightness and go wild with up/down arrows.
To convert to or from HSL, you can shift-click the colored box next to a
color property value.

Conversion was mostly done using https://github.com/atadams/Hex-to-HSL-Color
and some regular expressions.

Resources:
 - http://dev.opera.com/articles/view/color-in-opera-10-hsl-rgb-and-alpha-transparency/
 - http://en.wikipedia.org/wiki/HSL_and_HSV#Motivation
 - http://hslpicker.com