Add routing and some basic "Flux" architecture
Critic will be using a system inspired by Facebook's Flux
(http://facebook.github.io/flux/) to manage local data and handle
data flow through the application.

All these files use a require/exports (CommonJS) dependency system,
handled at runtime with the inluded require.js script (not to be
confused with the Require.JS project).

In the future we might want to use Webpack and make different bundles
for different parts of the application, to reduce the size of the
initial page load.