Add the Dashboard page
The Dashboard is the default page, and will display one or more lists of
reviews in different states (open, closed, dropped). The default state
is "open", but this can be customized by the `state` query parameter.
There is currently no way of reaching this query state from the UI.

When the Dashboard view component is first mounted in the DOM, it will
will request to fetch relevant data through a Review Action. On a
successful response, the data will be added to the ReviewStore.

The ReviewLists subview will be notified that something changed in the
ReviewStore, and will render the lists of reviews.