Add "Global Errors", for things like network errors
If a user or component makes a bad request to an API, or if a resources
could not be found, it's not an unrecoverable error. If there is no
network connection or if the server is down, we can show that in a
notifications bar below the top bar.

These errors will be handled in a separate store.